plot works again
