+++ interpreter/plot_addon/src/main/java/org/randoom/setlx/utilities/BarFrame.java
@@ -1,5 +1,7 @@
 package org.randoom.setlx.utilities;
 
+import org.jfree.chart.ChartPanel;
+import org.jfree.chart.JFreeChart;
 import org.jfree.chart.axis.CategoryAxis;
 import org.jfree.chart.axis.NumberAxis;
 import org.jfree.chart.plot.CategoryPlot;
@@ -26,6 +28,19 @@ public class BarFrame extends AbstractFrame {
 
     }
 
+    protected void redraw() {
+        if (chartCount != 0) {
+            jPanel.remove(chartPanel);
+        }
+        JFreeChart chart = new JFreeChart(title, JFreeChart.DEFAULT_TITLE_FONT, this.plot, true);
+
+        chartPanel = new ChartPanel(chart, true, true, true, true, true);
+
+        jPanel.add(chartPanel);
+
+        this.pack();
+        chartCount++;
+    }
     @Override
     public Graph addTextLabel(List<Double> coordinates, String text) {
         return null;
