+++ interpreter/plot_addon/src/main/java/org/randoom/setlx/utilities/DrawFrame.java
@@ -1,5 +1,7 @@
 package org.randoom.setlx.utilities;
 
+import org.jfree.chart.ChartPanel;
+import org.jfree.chart.JFreeChart;
 import org.jfree.chart.annotations.XYTextAnnotation;
 import org.jfree.chart.axis.NumberAxis;
 import org.jfree.chart.axis.ValueAxis;
@@ -65,7 +67,19 @@ public class DrawFrame extends AbstractFrame {
         yAxis = new NumberAxis("y");
         plot = null;
     }
+    protected void redraw() {
+        if (chartCount != 0) {
+            jPanel.remove(chartPanel);
+        }
+        JFreeChart chart = new JFreeChart(title, JFreeChart.DEFAULT_TITLE_FONT, this.plot, true);
 
+        chartPanel = new ChartPanel(chart, true, true, true, true, true);
+
+        jPanel.add(chartPanel);
+
+        this.pack();
+        chartCount++;
+    }
     protected void remakeFunctions() throws SetlException {
         plot = new XYPlot(new XYSeriesCollection(), (ValueAxis)xAxis, (ValueAxis)yAxis, new XYLineAndShapeRenderer());
         this.redraw();
